Split into speed and size optimised parts
